pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "oabot"
version = "0.1"
dependencies = [
    "mastodon.py==1.8.0",
    "requests==2.28",
    "pytest==7.1"
]
